+++ Controller/src/api/mod.rs
@@ -22,6 +22,7 @@ mod statistics;
 mod traffic_gen;
 mod ports;
 mod tables;
+mod histogram;
 
 mod helper;
 mod reset;
@@ -40,3 +41,4 @@ pub use ports::add_port;
 pub use reset::reset;
 pub use restart::restart;
 pub use config::config;
+pub use histogram::configure_histogram;