+++ Controller/src/api/server.rs
@@ -31,7 +31,7 @@ use utoipa::{openapi::security::{ApiKey, ApiKeyValue, SecurityScheme}, Modify, O
 use utoipa_swagger_ui::SwaggerUi;
 
 use tower_http::cors::{Any, CorsLayer};
-use crate::api::{configure_traffic_gen, online, ports, reset, statistics, stop_traffic_gen, traffic_gen, restart, add_port, config};
+use crate::api::{add_port, config, configure_histogram, configure_traffic_gen, online, ports, reset, restart, statistics, stop_traffic_gen, traffic_gen};
 
 
 use crate::api::helper::serve_static_files::{serve_index, static_path};
@@ -40,6 +40,7 @@ use crate::api::statistics::time_statistics;
 use crate::api::tables::tables;
 use crate::AppState;
 use crate::api::tables;
+use crate::api::histogram;
 
 use crate::core::traffic_gen_core::types::*;
 
@@ -54,7 +55,10 @@ use crate::core::traffic_gen_core::types::*;
         statistics::time_statistics,
         restart::restart,
         reset::reset,
-        ports::ports
+        ports::ports,
+        histogram::configure_histogram,
+        histogram::config,
+        online::online
     ),
     components(
         schemas(TrafficGenData,
@@ -156,6 +160,7 @@ pub async fn start_api_server(state: Arc<AppState>) {
         .route("/ports/arp", post(arp_reply))
         .route("/tables", get(tables))
         .route("/config", get(config))
+        .route("/histogram", get(histogram::config).post(configure_histogram))
         .layer(cors)
         .with_state(Arc::clone(&state));