+++ Controller/src/core/mod.rs
@@ -26,6 +26,9 @@ pub use frame_type_monitor::FrameTypeMonitor;
 pub mod duration_monitor;
 pub use duration_monitor::DurationMonitorTask;
 
+pub mod rtt_histogram_monitor;
+pub use rtt_histogram_monitor::HistogramMonitor;
+
 mod rate_monitor;
 pub mod statistics;
 pub mod traffic_gen;